add step to clean up the disc space to fix `No space left on device` (https://redirect.github.com/duckdb/duckdb/pull/17390)
This PR should fix
https://github.com/duckdblabs/duckdb-internal/issues/4753

There are two cases of this issue - when it runs out the space on Build
and on `temp_directory_enable_external_access.test`.

This fix should help to clean up at least 5GB additional space by
removing tools preinstalled in the GH runner (except for Python).

There are more preinstalled stuff to remove, so it this fix don't help,
we can remove more things.

Upstream-commit: https://github.com/duckdb/duckdb/commit/201df2b7c470589a91fa18ea59a98fcc3428d99c