[Dev] Add `QualifiedName::ParseComponents`, add input to the error messages (https://redirect.github.com/duckdb/duckdb/pull/17403)
I'm slightly abusing the `QualifiedName::Parse` method in iceberg
(https://github.com/duckdb/duckdb-iceberg/pull/224), which causes a
problem when the path contains multiple dots.
I'm adding `QualifiedName::ParseComponents` so we can make use of the
parsing logic safely.

One other thing that stood out to me was that an empty input is not
invalid, which feels wrong, but I preserved the behavior in this change.

Upstream-commit: https://github.com/duckdb/duckdb/commit/62d401e2da0812aa8b7a4546cb5b6d9d2dcd413e