Move all httplib code to HTTPUtil class (https://redirect.github.com/duckdb/duckdb/pull/17420)
This PR moves all `httplib` code to a separate `HTTPUtil` class. This
class lives in the `DBConfig`, and has virtual methods. The idea is that
this class implements support for all HTTP requests we need. This class
can then be overridden by another class (i.e. `httpfs`), similar to how
the encryption util works. This has a number of advantages:

* Limiting HTTP requests to a single class means it is easier for WASM
to share code with the main library
* Limiting HTTP requests to a single class means it is easier to swap
HTTP libraries (e.g. if we want to switch to using curl)
* Having HTTP requests in a single class means it is possible for other
extensions besides httpfs to issue HTTP(S) requests without themselves
bundling httplib/openssl

For now this is only the extension installation code. The idea is to
build on top of this in the `HTTPFS` extension and to move all of the
web requests to the `HTTPUtil` class in a subsequent PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ef5d5d8a38c35827edc906a290e291935c673126