unittest: improve detection of whether or not we can run `--force-restart` tests (https://redirect.github.com/duckdb/duckdb/pull/17419)
The `--force-restart` tests are useful to stress test database restarts
by restarting in between each statement - but we cannot always restart
since we have transient state (e.g. attached databases, prepared
statements, etc). Restarting will then lead to failing the test. For
tests that have this transient state we have been accumulating a lot of
`require skip_reload` flags. This is undesirable since this often breaks
CI only in the final stages of a PR (at which point these skips are
added).

This PR instead improves detection on when we can restart, allowing
almost all of the `require skip_reload` flags in tests to be removed. In
particular:

* We cannot restart if there are any attached databases (**new**)
* We cannot restart if there are any `TEMPORARY` structures (tables,
views, etc) (**new**)
* We cannot restart if a random seed has been set (**new**)
* We cannot restart if there are any prepared statements 
* We cannot restart if we are in the middle of a transaction

We introduce new syntax for specifying seeds in the tests:

```sql
set seed 0.42
```

This syntax sets the seed, and also disables the force restart tests.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ebecbe26caf95be6fa9c7ebb539a21d752544d48