On COPY TO/FROM check the format during binding. (https://redirect.github.com/duckdb/duckdb/pull/17381)
We previously had the format options hardcoded in the transformer. This
code changes this to attempt to get the proper COPY FROM/TO options from
the catalog during binding.

This change allows us to pick the correct COPY function on extensions,
based on the extension format, without having to explicitly set the
`(FORMAT ...)` option.

Upstream-commit: https://github.com/duckdb/duckdb/commit/691c73217b8570d5a4e9ca92d8aa6f46bab7213a