Force errors when trying lines as early as possible (https://redirect.github.com/duckdb/duckdb/pull/17427)
When detecting where a line starts in a CSV batch, if the file contains
quoted newlines, it could create a situation where the entire batch
needs to be processed before the system detects that it can't find a new
row. This drastically slows down CSV processing.

This PR changes that so errors are reported as early as possible.

Fix: https://github.com/duckdb/duckdb/issues/17362
Upstream-commit: https://github.com/duckdb/duckdb/commit/c5686ac15501ef64a4c855c1863ff18c530ebb03