Multi File Reader Rework (Part 19): Make `MultiFileReaderInterface` virtual, and move reading methods to the `BaseFileReader` (https://redirect.github.com/duckdb/duckdb/pull/17475)
This PR reworks the multi-file reader so that the
`MultiFileReaderInterface` is a virtual class that must be inherited
from instead of a class filled with static methods. This has a number of
advantages:

* The interface is well defined (virtual methods that must be
implemented)
* We can provide default implementations for methods that might not need
to be overloaded by all implementations
* This is extensible - we can add methods without needing to implement
them to all implementations by providing a default

The interface is obtained through a single static method that must be
defined on the interface object:

```cpp
static unique_ptr<MultiFileReaderInterface> InitializeInterface(ClientContext &context, MultiFileReader &reader, MultiFileList &file_list);
```

Several methods are also moved from the `MultiFileInterface` to the
`BaseFileReader`. This allows potentially different types of
`BaseFileReader` objects to be used within the same multi-file reader.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0671c7033902ad9571b88ca7782974fd5620545d