Emit dictionary vectors with unaligned start index (https://redirect.github.com/duckdb/duckdb/pull/17471)
Hi,

This PR enables emitting Dictionary Vectors even if the start index is
not aligned with `BITPACKING_ALGORITHM_GROUP_SIZE`. The logic is similar
to `ScanToFlatVector` in the sense that it might read extra values and,
at the end, it corrects the emitted selection vector.

When looking into the `Title` and `URL` column of Clickbench, I noticed
that the number of Dictionary Vectors emitted was far less than I
expected (~68% of the values are dictionary encoded in the URL column
for example). With this change, the number of emitted Dictionary Vector
is almost three times higher for the mentioned columns. Therefore, it
results in less memory usage and better utilization of
`TryAddCompressedGroups` in Aggregate HashTable.

Performance results for queries 33, 34, 36, 37, and 38 of Clickbench:

![emit_dict_comparison](https://github.com/user-attachments/assets/0a72c13f-33c2-4b11-8cf6-4ba01cd5d010)

The benchmark was ran on a MacBook Air M2, 16GB memory with 1 thread on
10% of the data.

The most notable performance gain was for query 33 with 10-12% speedup
as it contains no filtering and is a single column `GROUP BY`.

Are there any implicit behavior can be caused by this change? I have not
encountered any so far.
Please let me know if you'd like me to adjust anything!

Thanks

Upstream-commit: https://github.com/duckdb/duckdb/commit/d4d722c1c9d84c97d3518ddba484b2679257913c