Fix GeoParquet ExpressionColumnReader schema (https://redirect.github.com/duckdb/duckdb/pull/17481)
This PR fixes an issue where GeoParquet support broke due to changes
made to the parquet expression column reader.

Previously, if no schema was provided when creating a
ExpressionColumnReader, the schema of the child column reader was used,
but this no longer works as GEOMETRY columns now create a nested child
schema for the underlying blob column, meaning that the resulting
expression reader would have the same return type as the child, i.e.
BLOB, not GEOMETRY. The fix here is simply to pass the schema in the
ExpressionColumnReader constructor, and not copy the child schema.

However, there are other places where we add an expression reader on top
of an existing reader/schema, and create a new column schema. In these
cases its important that this new schema lives as long as the
ExpressionColumnReader, hence the ExpressionColumnReader now takes a
schema either as a const reference (if it already has a corresponding
schema, as is the case for GeoParquet), or as a "owning" unique pointer
(if its added on top of an existing schema afterwards).

Additionally, this PR also adds spatial to the nightly test CI to avoid
this breaking again unnoticed in the future.

Upstream-commit: https://github.com/duckdb/duckdb/commit/e92be6795f5683c4ef6873ee20ed7502f1d5d1aa