Add FileSystem::TryRemoveFile - that only removes a file if it exists (https://redirect.github.com/duckdb/duckdb/pull/17502)
Many places in the code have the following pattern:

```cpp
if (fs.FileExists(local_extension_path)) {
	fs.RemoveFile(local_extension_path);
}
```

This is both racy (what if the file is removed after the `FileExists`
call) and inefficient (need to do two syscalls/round-trips). This PR
provides a new method `TryRemoveFile` that only removes files if it
exists. Currently it still falls back to the original implementation,
but we can pull it through to the various file system layers in
follow-up PRs to avoid having to do two round-trips.

Upstream-commit: https://github.com/duckdb/duckdb/commit/507496bb12215a5b106bea2124886c7771f31294