Add OperatorFinalize callback to operators - which is called after a pipeline is finished (https://redirect.github.com/duckdb/duckdb/pull/17503)
This PR adds the `OperatorFinalize` callback. This callback is called
after execution of a pipeline is fully finished - similar to the Sink
`Finalize` call. This makes operators more powerful as we have a "final"
co-ordination point, which will allow for more flexible/powerful
operations to be implemented as (partially) streaming.

Upstream-commit: https://github.com/duckdb/duckdb/commit/af2d1d473bd3ad8dd18030b4953a3c93c1cf8107