Fix: Optional Probe Selection (https://redirect.github.com/duckdb/duckdb/pull/17505)
This fixes a bug introduced in
https://github.com/duckdb/duckdb/pull/17062, which occurred during hash
table probing when resolving hash collisions. The issue arose
specifically when NULL values on the build side were filtered out before
probing.

The PR also adds a test that can reproduces the issue. Since hash
collisions are relatively rare, the test uses a large join to increase
the likelihood of collisions. It takes about 40 seconds in debug mode
and 0.3 seconds in release mode.

Upstream-commit: https://github.com/duckdb/duckdb/commit/311f550603088192ec93da9b4cea47bc11a4d5a9