Output hashes in unittest and fix order (https://redirect.github.com/duckdb/duckdb/pull/17664)
Currently, no information about the hashes is shown when a unittest
fails because of a hash value mismatch, even though this information is
readily available.

This small PR aims to include this information in the output while also
making it less confusing: before, it printed an 'Expected result' header
without supplying any results (or rather, printing it in the previous
lines before the introductory header).

```
	PrintHeader("Expected result:");
	PrintLineSep();        <--- There should be a result here
	PrintHeader("Actual result:");
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/f87449e7b64f093b58bc92d47870648a22bdd178