ClientBufferManager wrapper to access the client context in the buffer manager (https://redirect.github.com/duckdb/duckdb/pull/17699)
EDIT: - [ ] move the client context back into the `ClientBufferManager`.

This PR lays the groundwork to make the client context accessible in the
`BufferManager`.

This is especially useful to collect metrics such as the allocated
memory of a query, or the time spent allocating memory.
The approach is similar to what we do for `ClientFileSystem` (which can
be used to collect file-system related metrics by passing the client
context).

I changed the `UndoBufferAllocator` to be based on the database
instance's buffer manager, directly, as the `ClientData` might not be
alive anymore when rolling back (other) transactions - @Mytherin maybe
you can sanity check this.

Another thing to note is that I've turned the `protected` functions of
the buffer manager into `public` functions.
Maybe I am missing something here, but I could not get the wrapper to
work for these functions otherwise, as I am not calling them from
`ClientBufferManager`, but its member `buffer_manager`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9a1fc9d085d56b45b14e5f1efd246695674e839f