Aggregation performance (https://redirect.github.com/duckdb/duckdb/pull/17718)
 * Increase `jemalloc` arena count
 * Slightly improve radix partitioning performance
* Fast path for appending small amounts of fixed-width data to
`TupleDataCollection`
 * Template `SelectionVector`s in various places
 * Inline hash functions for `int64_t`/`uint64_t`
* Don't align `TupleDataLayout` if there are no aggregate functions.
There's no need, and this saves up to 7 bytes per tuple
* Improve linear probing behavior: use some of the upper bits of the
hash (salt) to increment the offset to the next bucket. This reduces
collision clusters/chain length
 * Change load factor from 1.25 to 1.5 to reduce collisions further
* Lower initial `sink_capacity` to `32768`. We already implemented
HLL-based adaptation, so the HT size will increase later if the data
requires it
* Get the next hash partition index to scan using an atomic instead of
with a lock

Together, this improves the performance of simple `SELECT COUNT(DISTINCT
<col>) FROM <table>;` queries on integers by ~11% (strings by ~5%), on
my laptop. Other kinds of grouped aggregations should also improve, but
likely to a lesser extent.

Upstream-commit: https://github.com/duckdb/duckdb/commit/896982b90fc6823b02161e430548a9bc1845c88a