Adding additional authenticated data for encryption (https://redirect.github.com/duckdb/duckdb/pull/17508)
This PR adds the possibility to include Additional Authenticated Data
(AAD) data for encryption. For some ciphers like AES GCM, AAD can be
used to make sure that the ciphertext is not tampered with. This is e.g.
used in the Parquet encryption spec, and we might want to use it for our
own encryption.

Upstream-commit: https://github.com/duckdb/duckdb/commit/32009cdbaa8e32a01fc9ad083a375e60885a85f6