[Indexes] Buffer-managed indexes part 1: segment handles (https://redirect.github.com/duckdb/duckdb/pull/17758)
Currently, indexes are not buffer-managed, i.e., the buffer manager
never unpins index memory. That means indexes fill up memory, and we
cannot unpin them if we need extra memory - even if we currently don't
use any index memory. For persistent databases, detaching and
re-attaching the database frees up the index memory (temporarily), as
indexes are loaded lazily.

But, of course, that workaround only works for persistent databases and
is not ideal.

Thus, this PR is the first in a series of planned PRs to make indexes
buffer-managed.

The main change is the introduction of a reader count for each
`FixedSizeBuffer` and a `SegmentHandle`, which wraps the memory segments
of indexes. These handles track the lifetime of a segment of index
memory and handle the reader count of their `FixedSizeBuffer`.

Currently, segment handles exist in addition to the previous
functionality. **We do not yet have buffer-manage indexes.** The idea is
to incrementally rewrite the code to use segment handles and,
eventually, remove the old functions. Then, we can push an eviction
strategy for fixed-size buffers.

We already use the `SegmentHandle` inside the `FixesSizeBuffer` and
`FixedSizeAllocator`. Additionally, I've added a`NodeHandle` wrapper and
rewritten the `Node256Leaf` code to use these node handles.

cc @Maxxen 

Upstream-commit: https://github.com/duckdb/duckdb/commit/926bd2531d4f020b1f4e296af17cc5b668f0aae3