Replace string for const data ptr in encryption api (https://redirect.github.com/duckdb/duckdb/pull/17825)
The encryption API took a string for the key as input, meaning that we
are forced to cast bytes to a string (which is not safe). This PR
changes the API slightly such that we don't have to cast bytes to a
string before using encryption functions and is necessary to avoid
storing keys in strings.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0836b6d96fb392d68589091159ba27c23c923540