Make CTE Materialization the Default Instead of Inlining (https://redirect.github.com/duckdb/duckdb/pull/17459)
Until now, DuckDB's default behavior for CTEs was to inline them, which
means that
the CTE is replaced with its definition in the main query. However,
inlining can not
only lead to performance issues, especially when the CTE is used
multiple times in
the main query, but also to incorrect results in some cases.

With the introduction of the `WITH ... AS [NOT] MATERIALIZED` syntax in
https://github.com/duckdb/duckdb/pull/7126,
users can specify that a CTE should be materialized, meaning that it
will be
computed once and stored in memory for the duration of the main query.
This can improve performance and ensure correct results when the CTE is
used multiple times.

This PR is a continuation of that work. While we still want to be able
to inline CTEs in some cases,
it should not be the default behavior anymore. More importantly, this
should not be the responsibility
of the `Binder`, but rather an `Optimizer` decision.

For CTEs with `MATERIALIZED` or `NOT MATERIALIZED` specified, nothing
changes.
If a CTE is marked as `MATERIALIZED`, it will be materialized, and if it
is marked
as `NOT MATERIALIZED`, it will be inlined—regardless of the possibly
questionable
performance or correctness implications.

For CTEs without a `MATERIALIZED` or `NOT MATERIALIZED` clause, the
`Optimizer` will
decide whether to inline or materialize the CTE.

The rules for this decision are currently a work in progress.

Upstream-commit: https://github.com/duckdb/duckdb/commit/e00dfa4b49fbd2374efe5fe703b9fa643cf78500

---

**Performance**

@neumannt released a solution of the Advent of Code 2024 using SQL:
https://github.com/neumannt/aoc24
All solutions rely heavily on CTEs, but do not use explicit
MATERIALIZATION hints.
I use his queries as a benchmark—using the full AoC input:

| DAY | Speedup with this PR vs. DuckDB main |
|-----|---------------------------------------|
| 1   |  2.3x                                 |
| 2   |  1.7x                                 |
| 3   |  1.6x                                 |
| 4   |  2.7x                                 |
| 5   | 99.3x                                 |
| 6   | 1810x (now ~18 sec, main ~9h)               |
| 7   | 11.8x                                 |
| 8   |  1.2x                                 |
| 9   |156.2x                                 |
| 10  |  1.7x                                 |
| 11  |  1.2x                                 |
| 12  |  2.3x                                 |
| 13  |  3.2x                                 |
| 14  |  308x (now ~1.5s, main ~8 min)               |
| 15  | > 10x (now ~2 min, main timeout)       |
| 16  | TBD                                   |
| 17  | Not supported                         |
| 18  | > 73x (now ~8.5 sec, main timeout)     |
| 19  |  2.2x                                 |
| 20  | 28x (now ~8 sec, main ~3 min 50s)           |
| 21  | 24.2x                                 |
| 22  | Not supported                         |
| 23  | 188x (now ~260ms, main ~48 sec)               |
| 24  | Not supported                         |
| 25  | > 2245x (now ~250ms, main timeout)     |

I think these numbers speak for themselves. All queries experienced a
speedup, and
the most problematic queries are now fast enough to run in a reasonable
time frame.
Speedups range from 1.2x to over 2000x, which—I think—is a significant
improvement.

---

`logical_operator_deep_copy` is a new visitor that allows us to _"deep
copy"_ a logical operator tree.
While we can already get a 1:1 copy of a logical operator using `Copy`,
we can not use that
copy to immediately replace a CTE reference in the main query. This is
because the copy
has the same `table_index`es as the original tree, which breaks some
sanity checks.
To fix this, we traverse the tree and assign new `table_index`es to each
operator.

---

TODO:
- [x] Improve the rules for inlining vs. materializing CTEs
- [x] Fix the unit tests for `WITH RECURSIVE` queries, that are
currently failing because they relied on the inlining behavior
- [x] Clean up some rough edges in the code
- [x] Add more tests for the new `Optimizer` behavior

---

Fix: https://redirect.github.com/duckdb/duckdb/pull/17456
Fix: https://github.com/duckdb/duckdb/issues/12875
Fix: https://github.com/duckdb/duckdb/issues/15416
Fix: https://github.com/duckdb/duckdb/issues/16891
Fix: https://github.com/duckdb/duckdb/issues/9940
Fix: https://github.com/duckdb/duckdb/issues/14244