Leverage `VectorType` in `ColumnDataCollection` (https://redirect.github.com/duckdb/duckdb/pull/17881)
When materializing data in columnar format in the
`ColumnDataCollection`, we (almost) always go to the `FLAT`
representation, i.e., `CONSTANT` and `DICTIONARY` `Vector`s are
flattened. For most data types this is OK, but for strings, this can
cause the size of the intermediates to increase drastically.

Keeping the `VectorType` the same would be ideal, but when we scan, we
expect the `Vector`s to be `FLAT`, so this would be a big (and
bug-prone) effort, so it's not really feasible (for now). Instead, this
PR copies over just the unique strings, and creates a `FLAT` `Vector` of
strings that point into it. This solves most of the problem while
keeping a very low code footprint.

This optimization will trigger always for `CONSTANT` `Vector`s, but for
`DICTIONARY` `Vector`s only if the dictionary size is less than half of
our `STANDARD_VECTOR_SIZE`. The dictionary is copied once per `Vector`.
This can be improved in the future by copying the dictionary just once.
This is totally feasible, but more effort to implement. What I've
implemented here are the first steps in that direction, but I've chosen
to keep this PR simple for now, as I think this solves 80% of the
problem with just 20% of the effort.

I did a quick benchmark by copying the `l_shipinstruct` column from
`lineitem` from TPC-H at SF100 from Parquet to Parquet, and this reduced
the peak RSS from 5.1 GB to 3.5 GB. The speed was pretty much the same.
I can come up with a lot more degenerate examples where memory usage
would be reduced by orders of magnitude.

Upstream-commit: https://github.com/duckdb/duckdb/commit/fd1b726c1cea7ef33ca2334bc19a6f76cb317bb1