Block based encryption (https://redirect.github.com/duckdb/duckdb/pull/17275)
This PR adds block-based encryption. In block-based encryption all
blocks - except the main database header - are encrypted. To ensure that
encryption is randomised, each block is prefixed with a 12-byte nonce.
For the encryption mode `AES-GCM` the nonce is followed by a 16-byte
tag, which is calculated after encrypting the actual data and checksum
of the corresponding block. Upon decryption, this tag is calculated and
checked against the stored tag (pretty much like how a checksum works).

**Encryption Algorithm support**
For now we only support AES-GCM with a fixed key length of 32 bytes
(`aes-gcm-256`) in Mbed TLS. This is a randomised encryption algorithm
that also performs authentication by means of a calculated tag.

## Usage

**Encrypt an existing database**

```
ATTACH 'existing_unencrypted.duckdb' as unencrypted;
ATTACH 'encrypted.duckdb' AS encrypted (ENCRYPTION_KEY 'asdf');
COPY FROM DATABASE unencrypted TO encrypted;
```

**Decrypt an existing database**
```
ATTACH 'encrypted.duckdb' AS encrypted (ENCRYPTION_KEY 'asdf');
ATTACH 'new_unencrypted.duckdb' as unencrypted;
COPY FROM DATABASE encrypted TO unencrypted;
```

**Reencrypt an existing database**
```
ATTACH 'existing_encrypted.duckdb' AS existing_encrypted (ENCRYPTION_KEY 'asdf');
ATTACH 'new_encrypted.duckdb' AS new_encrypted (ENCRYPTION_KEY 'xxxx');
COPY FROM DATABASE existing_encrypted TO new_encrypted;
```

**Write to an encrypted database**
```
ATTACH 'encrypted.duckdb' AS encrypted (ENCRYPTION_KEY 'asdf');
CREATE OR REPLACE TABLE encrypted.tbl AS SELECT * FROM range(10) t(i);
SELECT SUM(i) FROM encrypted.tbl;
```

## Performance

The benchmarks are performed on a Mac M2 with 16GB RAM. All queries are
performed with TPC-H scale factor 10, unless specified otherwise.

| Operation | Time |
|:---|:---|
| `COPY FROM DATABASE unencrypted TO new_unencrypted` | 10.639 s |
| `COPY FROM DATABASE unencrypted TO encrypted` | 12.688 s |
| `COPY FROM DATABASE encrypted TO unencrypted` | 12.382 s |
| `COPY FROM DATABASE encrypted TO new_encrypted`  | 14.528 s |
| `SELECT SUM(l_suppkey) FROM lineitem` unencrypted | 0.067 s |
| `SELECT SUM(l_suppkey) FROM lineitem` encrypted | 0.252 s |
| (**SF1**) `SELECT SUM(l_suppkey) FROM lineitem` unencrypted | 0.009 s
|
| (**SF1**) `SELECT SUM(l_suppkey) FROM lineitem` encrypted | 0.027s |

Encryption incurs significant overhead. With using Mbed TLS, for a
simple aggregate in an encrypted file the overhead is around ~3x for
TPC-H SF1 Lineitem table, and around ~4x for SF10.

## Tests
For all above use-cases, I've added corresponding SQLLogic tests.

**Other tests**
To ensure that the database is actually encrypted, we can (i) calculate
the entropy with `ent encrypted.duckdb` (ii) search for existing strings
that we know, e.g. `strings encrypted.duckdb | grep "AIRMAIL"`, or (iii)
inspect the file ourselves with `hexdump`. For an encrypted database in
TPC-H SF10, the entropy is 8 bits per byte, which means that the file is
(almost) completely random. Also, existing strings such as `AIRMAIL`
cannot be found. For an unencrypted TPC-H SF10 file the calculated
entropy is around 7.7 - which is also pretty high - but this is because
of compression.

## Internals

**Key management**
A user can input a key of any length, and this will be reduced or
expanded to a 32-byte key. This is done by a key derivation function
(kdf). For hashing the password we use `SHA256`, which we could replace
with a more secure kdf. To replace the kdf however we will have to
expand the (stripped) Mbed TLS library. The derived key is managed by
the `EncryptionKeyManager`, and the derived key id is stored in the
`EncryptionOptions`. As an extra security measurement, we lock the
encryption key with mlock/virtuallock to avoid that the key is ever
swapped to disk.

**Wrong Key Detection**
To distinguish between a checksum failure and a wrong key, we store a
`canary` in the database header. The plaintext of this canary
(`DUCKKEY`) is known, and the canary will be encrypted along with the
database. If we then later read an encrypted database, we first decrypt
the canary and check whether the decrypted canary value is correct. If
the used key is not correct, the database will throw an error.

**Metadata**
If the database is encrypted, the first flag of the main database header
is set to 1. Additionally, we thus store a (i) canary and (ii) 8-bytes
of additional metadata. The first byte of the additional metadata
indicates the kdf used, the second byte is for the usage of AAD, the
third byte indicates the cipher, and the last 4 bytes identify the
`key_length`. This additional metadata is not used yet, but we will need
this later when we start supporting multiple kdfs, key lengths and
ciphers (such as AES CTR and AES CBC).

Upstream-commit: https://github.com/duckdb/duckdb/commit/37736e216e12100e498c5bc4d972afa643c30e0e