Add v1.3-ossivalis to Cross version workflow (https://redirect.github.com/duckdb/duckdb/pull/17906)
I was adding a `SUMMARIZE_FAILURES` env variable into the steps running
unit tests without python runner and found that `v1.3-ossivalis` is not
present in the versions matrix of this workflow. Probably we'd like to
have it in test matrix of `CrossVersion.yml`.

This PR adds `v1.3-ossivalis` into the test matrix, adds a concurrency
group ~~and read permission (which
[recommended](https://github.com/actions/checkout?tab=readme-ov-file#recommended-permissions)
to set when using action/checkout)~~.

Upstream-commit: https://github.com/duckdb/duckdb/commit/23f4de5996157fe435b6160e6bb88b5821b1fcab