Unit Tester Configuration (https://redirect.github.com/duckdb/duckdb/pull/17972)
This PR reworks the unittester configuration by introducing a new class
- `TestConfiguration` - that centralizes the (previously scattered) ways
of configuring the unittester. The following options are supported:

| Name | Description |

|------------------------|--------------------------------------------------------------|
| description | Config description |
| initial_db | Initial database path |
| max_threads | Max threads to use during tests |
| checkpoint_wal_size | Size in bytes after which to trigger automatic
checkpointing |
| checkpoint_on_shutdown | Whether or not to checkpoint on database
shutdown |
| force_restart | Force restart the database between runs |
| summarize_failures | Print a summary of all test failures after
running |
| test_memory_leaks | Run memory leak tests |
| verify_vector | Run vector verification for a specific vector type |
| debug_initialize | Initialize buffers with all 0 or all 1 |
| init_script | Script to execute on init |
| on_init | Script to execute on init |


### Configuration

There are three ways of setting options:

**Command Line.**, replacing `_` with `-`, e.g.:

```
# booleans
unittest --checkpoint-on-shutdown
unittest --no-checkpoint-on-shutdown
# other types
unittest --checkpoint-wal-size 100000
```

**Environment Variable**, prefixed with `DUCKDB_TEST_` and upper-cased,
e.g.:

```
DUCKDB_TEST_CHECKPOINT_ON_SHUTDOWN=1 unittest
```

**Configuration File**, a JSON file that can be passed in using
`--test-config`, e.g.:

```json
{
   "checkpoint_wal_size": "42",
   "checkpoint_on_shutdown": "false"
}
```

```
unittest --test-config my_config.json
```

### on_init / init_script

In addition to porting existing options, we introduce the `on_init`
option. This option can be used to execute a set of SQL queries before
running a test. For example:

```
unittest --on-init "ATTACH 'my_file.db'; USE my_file"
```

The `--init-script` parameter can be used to instead read the scripts
from a file.

This can be used to easily execute tests on different storage back-ends,
or pass configuration.

### Summarize Failures

This PR also extends the summarize failures. It is now always enabled by
default in CI (by checking for the presence of the `CI` variable). In
addition, the new `description` field can be used to provide an extended
test description. When summarize failures is enabled, the description is
printed above the failing tests, allowing the CI to provide context on
how to reproduce/what to do about a test failure.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2c08be48592d928fef220c3f418dd5699a88a2ff