Avoid adding commands read from a file to the shell history (https://redirect.github.com/duckdb/duckdb/pull/18057)
Previously when executing files, through e.g. `.read file.sql`, the
content of those files would be added to the shell history.

This is rather confusing and can fill up the history with large
statements quickly. It also does not match the behavior of other shells
(e.g. sqlite or postgres).

This PR makes it so that commands that are read from a file are not
added to the shell history.

Upstream-commit: https://github.com/duckdb/duckdb/commit/5639983f7d650a57f2217f832334949d8ba920c3